+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Added the Marburg 2026 HTML exercises for sessions 03 through 10 to the
   workshop resource catalogue.
+* Added the Marburg 2026 PDF slides for sessions 01 through 10 to the workshop
+  resource catalogue.
 
 # SfL 1.0.1